files_new fix

+++ bin/3.sh
@@ -1,3 +1,5 @@
 . etc/urf.conf
-mv var/files_new var/files_old
+if [ -f var/files_new ]; then
+   mv var/files_new var/files_old 
+fi
 find $picsdir -type f | sort | sed "s@$picsdir/@@g" > var/files_new
